feat(riscv): more isel rules for i32 (#964)
this is a complement to Sarah's recent patch, it's sort of a mechanical
addition of a bunch of isel rules for i32, so we can hopefully lower
more real applications. I hope it's boring code! some of it such as the
32-bit `bitreverse` is a little bit exciting but I've tested it fairly
thoroughly and also the refinement proof goes.

one nasty thing here is a little patch to the Veir->MIR printer, which
was emitting incorrect MIR code for 32->64 conversions, yikes! these
need to be zero extensions.

this patch has been fuzzed and seems solid.

+++ Test/Passes/InstructionSelection/RISCV64/srem_invalid.mlir
@@ -1,10 +1,10 @@
 // RUN: veir-opt %s -p=isel-riscv64 | filecheck %s
 
 "builtin.module"() ({
-    "func.func"()  <{function_type = (i32, i32) -> ()}> ({
-    ^bb0(%a: i32, %b: i32):
-        %srem = "llvm.srem"(%a, %b) : (i32, i32) -> i32
-        // CHECK: %{{.*}} = "llvm.srem"(%{{.*}}, %{{.*}}) : (i32, i32) -> i32
+    "func.func"()  <{function_type = (i16, i16) -> ()}> ({
+    ^bb0(%a: i16, %b: i16):
+        %srem = "llvm.srem"(%a, %b) : (i16, i16) -> i16
+        // CHECK: %{{.*}} = "llvm.srem"(%{{.*}}, %{{.*}}) : (i16, i16) -> i16
 
         "func.return"() : () -> ()
     }) : () -> ()
